new project with a webhook for receiving callbacks (an alternative to the change reader)
