+++ AlertIntegrationDemo/DtoAlertChanges/BookmarkContainer.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace BiQ.AlertIntegrationDemo.DtoAlertChanges
+{
+    public sealed class BookmarkContainer
+    {
+        public DateTimeOffset Bookmark { get; set; }
+    }
+}
