+++ AlertIntegrationDemo/JsonConverter.cs
@@ -2,7 +2,6 @@
 using Newtonsoft.Json.Linq;
 using Newtonsoft.Json.Serialization;
 using System;
-using System.IO;
 using System.Linq;
 
 namespace BiQ.AlertIntegrationDemo
