+++ ChangeCallbackReceiver/ChangeCallbackReceiver.csproj
@@ -1,0 +1,17 @@
+<Project Sdk="Microsoft.NET.Sdk.Web">
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <InvariantGlobalization>true</InvariantGlobalization>
+    <PublishAot>true</PublishAot>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\AlertIntegrationDemo\AlertIntegrationDemo.csproj" />
+    <ProjectReference Include="..\ChangeReader\ChangeReader.csproj" />
+    <ProjectReference Include="..\CustomerSystem\CustomerSystem.csproj" />
+  </ItemGroup>
+
+</Project>
